combine the tests
