--- End2End/HTTP/src/main/functions/local.settings.json
@@ -1,2 +1,0 @@
-{
-}
