--- End2End/Queue/src/main/functions/host.json
@@ -1,2 +1,0 @@
-{
-}
