--- End2End/Timer/src/main/functions/host.json
@@ -1,2 +1,0 @@
-{
-}
