--- End2End/Timer/src/main/functions/local.settings.json
@@ -1,7 +1,0 @@
-{
-    "IsEncrypted": false,
-    "Values": {
-      "AzureWebJobsStorage": "",
-      "AzureWebJobsDashboard": ""
-    }
-  }
