+++ End2End/functest/pom.xml
@@ -21,7 +21,7 @@
             <dependency>
                 <groupId>com.microsoft.azure.functions</groupId>
                 <artifactId>azure-functions-java-library</artifactId>
-                <version>1.0.0-beta-4</version>
+                <version>1.0.0-beta-5</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
